feat: add util function to prepare iso code from inputs. Update buy sell order submit hook
